requirements.txt
streamlit==1.52.2
pandas
numpy
scikit-learn
sentence-transformers
pdfminer.six
pymongo
matplotlib
python-dotenv
